feat(frontend): improve sub-agent observability
