+++ frontend/src/components/workspace/messages/message-steps.tsx
@@ -1,0 +1,333 @@
+import type { Message } from "@langchain/langgraph-sdk";
+import {
+  BookOpenTextIcon,
+  FolderOpenIcon,
+  GlobeIcon,
+  LightbulbIcon,
+  ListTodoIcon,
+  MessageCircleQuestionMarkIcon,
+  NotebookPenIcon,
+  SearchIcon,
+  SquareTerminalIcon,
+  WrenchIcon,
+} from "lucide-react";
+
+import {
+  ChainOfThoughtSearchResult,
+  ChainOfThoughtSearchResults,
+  ChainOfThoughtStep,
+} from "@/components/ai-elements/chain-of-thought";
+import { CodeBlock } from "@/components/ai-elements/code-block";
+import { useI18n } from "@/core/i18n/hooks";
+import {
+  extractReasoningContentFromMessage,
+  findToolCallResult,
+} from "@/core/messages/utils";
+import { extractTitleFromMarkdown } from "@/core/utils/markdown";
+
+import { useArtifacts } from "../artifacts";
+import { Tooltip } from "../tooltip";
+
+export interface CoTReasoningStep {
+  id?: string;
+  messageId?: string;
+  type: "reasoning";
+  reasoning: string | null;
+}
+
+export interface CoTToolCallStep {
+  id?: string;
+  messageId?: string;
+  type: "toolCall";
+  name: string;
+  args: Record<string, unknown>;
+  result?: string;
+}
+
+export type CoTStep = CoTReasoningStep | CoTToolCallStep;
+
+export function convertToSteps(messages: Message[]): CoTStep[] {
+  const steps: CoTStep[] = [];
+  for (const message of messages) {
+    if (message.type === "ai") {
+      const reasoning = extractReasoningContentFromMessage(message);
+      if (reasoning) {
+        const step: CoTReasoningStep = {
+          id: message.id,
+          messageId: message.id,
+          type: "reasoning",
+          reasoning: extractReasoningContentFromMessage(message),
+        };
+        steps.push(step);
+      }
+      for (const tool_call of message.tool_calls ?? []) {
+        if (tool_call.name === "task") {
+          continue;
+        }
+        const step: CoTToolCallStep = {
+          id: tool_call.id,
+          messageId: message.id,
+          type: "toolCall",
+          name: tool_call.name,
+          args: tool_call.args,
+        };
+        const toolCallId = tool_call.id;
+        if (toolCallId) {
+          const toolCallResult = findToolCallResult(toolCallId, messages);
+          if (toolCallResult) {
+            try {
+              const json = JSON.parse(toolCallResult);
+              step.result = json;
+            } catch {
+              step.result = toolCallResult;
+            }
+          }
+        }
+        steps.push(step);
+      }
+    }
+  }
+  return steps;
+}
+
+export function ToolCall({
+  id,
+  messageId,
+  name,
+  args,
+  result,
+  isLast = false,
+  isLoading = false,
+}: {
+  id?: string;
+  messageId?: string;
+  name: string;
+  args: Record<string, unknown>;
+  result?: string | Record<string, unknown>;
+  isLast?: boolean;
+  isLoading?: boolean;
+}) {
+  const { t } = useI18n();
+  const { setOpen, autoOpen, autoSelect, selectedArtifact, select } =
+    useArtifacts();
+
+  if (name === "web_search") {
+    let label: React.ReactNode = t.toolCalls.searchForRelatedInfo;
+    if (typeof args.query === "string") {
+      label = t.toolCalls.searchOnWebFor(args.query);
+    }
+    return (
+      <ChainOfThoughtStep key={id} label={label} icon={SearchIcon}>
+        {Array.isArray(result) && (
+          <ChainOfThoughtSearchResults>
+            {result.map((item) => (
+              <ChainOfThoughtSearchResult key={item.url}>
+                <a href={item.url} target="_blank" rel="noopener noreferrer">
+                  {item.title}
+                </a>
+              </ChainOfThoughtSearchResult>
+            ))}
+          </ChainOfThoughtSearchResults>
+        )}
+      </ChainOfThoughtStep>
+    );
+  } else if (name === "image_search") {
+    let label: React.ReactNode = t.toolCalls.searchForRelatedImages;
+    if (typeof args.query === "string") {
+      label = t.toolCalls.searchForRelatedImagesFor(args.query);
+    }
+    const results = (
+      result as {
+        results: {
+          source_url: string;
+          thumbnail_url: string;
+          image_url: string;
+          title: string;
+        }[];
+      }
+    )?.results;
+    return (
+      <ChainOfThoughtStep key={id} label={label} icon={SearchIcon}>
+        {Array.isArray(results) && (
+          <ChainOfThoughtSearchResults>
+            {Array.isArray(results) &&
+              results.map((item) => (
+                <Tooltip key={item.image_url} content={item.title}>
+                  <a
+                    className="size-24 overflow-hidden rounded-lg object-cover"
+                    href={item.source_url}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                  >
+                    <div className="bg-accent size-24">
+                      <img
+                        className="size-full object-cover"
+                        src={item.thumbnail_url}
+                        alt={item.title}
+                        width={100}
+                        height={100}
+                      />
+                    </div>
+                  </a>
+                </Tooltip>
+              ))}
+          </ChainOfThoughtSearchResults>
+        )}
+      </ChainOfThoughtStep>
+    );
+  } else if (name === "web_fetch") {
+    const url = (args as { url: string })?.url;
+    let title = url;
+    if (typeof result === "string") {
+      const potentialTitle = extractTitleFromMarkdown(result);
+      if (potentialTitle && potentialTitle.toLowerCase() !== "untitled") {
+        title = potentialTitle;
+      }
+    }
+    return (
+      <ChainOfThoughtStep
+        key={id}
+        label={t.toolCalls.viewWebPage}
+        icon={GlobeIcon}
+      >
+        <ChainOfThoughtSearchResult>
+          {url && (
+            <a
+              href={url}
+              target="_blank"
+              rel="noopener noreferrer"
+              className="cursor-pointer"
+            >
+              {title}
+            </a>
+          )}
+        </ChainOfThoughtSearchResult>
+      </ChainOfThoughtStep>
+    );
+  } else if (name === "ls") {
+    let description: string | undefined = (args as { description: string })
+      ?.description;
+    if (!description) {
+      description = t.toolCalls.listFolder;
+    }
+    const path: string | undefined = (args as { path: string })?.path;
+    return (
+      <ChainOfThoughtStep key={id} label={description} icon={FolderOpenIcon}>
+        {path && (
+          <ChainOfThoughtSearchResult className="cursor-pointer">
+            {path}
+          </ChainOfThoughtSearchResult>
+        )}
+      </ChainOfThoughtStep>
+    );
+  } else if (name === "read_file") {
+    let description: string | undefined = (args as { description: string })
+      ?.description;
+    if (!description) {
+      description = t.toolCalls.readFile;
+    }
+    const { path } = args as { path: string; content: string };
+    return (
+      <ChainOfThoughtStep key={id} label={description} icon={BookOpenTextIcon}>
+        {path && (
+          <ChainOfThoughtSearchResult className="cursor-pointer">
+            {path}
+          </ChainOfThoughtSearchResult>
+        )}
+      </ChainOfThoughtStep>
+    );
+  } else if (name === "write_file" || name === "str_replace") {
+    let description: string | undefined = (args as { description: string })
+      ?.description;
+    if (!description) {
+      description = t.toolCalls.writeFile;
+    }
+    const path: string | undefined = (args as { path: string })?.path;
+    if (isLoading && isLast && autoOpen && autoSelect && path && !result) {
+      setTimeout(() => {
+        const url = new URL(
+          `write-file:${path}?message_id=${messageId}&tool_call_id=${id}`,
+        ).toString();
+        if (selectedArtifact === url) {
+          return;
+        }
+        select(url, true);
+        setOpen(true);
+      }, 100);
+    }
+
+    return (
+      <ChainOfThoughtStep
+        key={id}
+        className="cursor-pointer"
+        label={description}
+        icon={NotebookPenIcon}
+        onClick={() => {
+          select(
+            new URL(
+              `write-file:${path}?message_id=${messageId}&tool_call_id=${id}`,
+            ).toString(),
+          );
+          setOpen(true);
+        }}
+      >
+        {path && (
+          <ChainOfThoughtSearchResult className="cursor-pointer">
+            {path}
+          </ChainOfThoughtSearchResult>
+        )}
+      </ChainOfThoughtStep>
+    );
+  } else if (name === "bash") {
+    const description: string | undefined = (args as { description: string })
+      ?.description;
+    if (!description) {
+      return t.toolCalls.executeCommand;
+    }
+    const command: string | undefined = (args as { command: string })?.command;
+    return (
+      <ChainOfThoughtStep
+        key={id}
+        label={description}
+        icon={SquareTerminalIcon}
+      >
+        {command && (
+          <CodeBlock
+            className="mx-0 cursor-pointer border-none px-0"
+            showLineNumbers={false}
+            language="bash"
+            code={command}
+          />
+        )}
+      </ChainOfThoughtStep>
+    );
+  } else if (name === "ask_clarification") {
+    return (
+      <ChainOfThoughtStep
+        key={id}
+        label={t.toolCalls.needYourHelp}
+        icon={MessageCircleQuestionMarkIcon}
+      ></ChainOfThoughtStep>
+    );
+  } else if (name === "write_todos") {
+    return (
+      <ChainOfThoughtStep
+        key={id}
+        label={t.toolCalls.writeTodos}
+        icon={ListTodoIcon}
+      ></ChainOfThoughtStep>
+    );
+  } else {
+    const description: string | undefined = (args as { description: string })
+      ?.description;
+    return (
+      <ChainOfThoughtStep
+        key={id}
+        label={description ?? t.toolCalls.useTool(name)}
+        icon={WrenchIcon}
+      ></ChainOfThoughtStep>
+    );
+  }
+}
+
+export { extractReasoningContentFromMessage, LightbulbIcon };
