+++ frontend/src/core/i18n/locales/types.ts
@@ -224,6 +224,10 @@ export interface Translations {
     in_progress: string;
     completed: string;
     failed: string;
+    taskAnalysis: string;
+    analyzingTask: string;
+    decomposedInto: (count: number) => string;
+    subagentType: (type: string) => string;
   };
 
   // Token Usage
