+++ frontend/src/core/tasks/context.tsx
@@ -1,3 +1,4 @@
+import type { AIMessage } from "@langchain/langgraph-sdk";
 import { createContext, useCallback, useContext, useState } from "react";
 
 import type { Subtask } from "./types";
@@ -41,10 +42,25 @@ export function useSubtask(id: string) {
 export function useUpdateSubtask() {
   const { tasks, setTasks } = useSubtaskContext();
   const updateSubtask = useCallback(
-    (task: Partial<Subtask> & { id: string }) => {
-      tasks[task.id] = { ...tasks[task.id], ...task } as Subtask;
-      if (task.latestMessage) {
+    // _appendMessage is an internal protocol: when present, dedup and append to
+    // messageHistory, then trigger a React state update. Render-path callers
+    // (message-list.tsx) omit _appendMessage, so their calls only mutate the
+    // tasks record in-place without triggering setTasks (avoids infinite loops).
+    (task: Partial<Subtask> & { id: string; _appendMessage?: AIMessage }) => {
+      const { _appendMessage: appendMsg, ...rest } = task;
+      const current = tasks[task.id];
+      const updated = { ...current, ...rest } as Subtask;
+
+      if (appendMsg) {
+        const history = [...(current?.messageHistory ?? [])];
+        if (!history.some((m) => m.id === appendMsg.id)) {
+          history.push(appendMsg);
+        }
+        updated.messageHistory = history;
+        tasks[task.id] = updated;
         setTasks({ ...tasks });
+      } else {
+        tasks[task.id] = updated;
       }
     },
     [tasks, setTasks],
